4 more shell commands + better paging
Only the first 4mb are now identity mapped.
The kernel is mapped to 0xC0000000
Added the add, multiply, divide, and subtract commands for terminal.

+++ kernel/arch/i386/pmm.c
@@ -48,7 +48,7 @@ uint32_t first_frame() {
 
 uint32_t alloc_frame() {
 	uint32_t idx = first_frame();
-	if (idx != -1) {
+	if (idx != (uint32_t)-1) {
 		use_bit(frames, idx);
 		free_frames--;
 		return idx * 0x1000;
